+++ R/method-ggplot-add.R
@@ -86,7 +86,7 @@ ggplot_add.geom_range <- function(object, plot, object_name) {
 }
 
 ##' @method ggplot_add layout_ggtree
-##' @importFrom ggplot2 expand_scale
+##' @importFrom ggplot2 expansion
 ##' @export
 ggplot_add.layout_ggtree <- function(object, plot, object_name) {
     if(object$layout == 'fan') {
@@ -102,7 +102,7 @@ ggplot_add.layout_ggtree <- function(object, plot, object_name) {
         ## refer to: https://github.com/GuangchuangYu/ggtree/issues/6
         ## and also have some space for tree scale (legend)
         obj <- list(coord_polar(theta='y', start=-pi/2, -1, clip = 'off'),
-                    scale_y_continuous(limits = c(0, NA), expand = expand_scale(0, 0.6))
+                    scale_y_continuous(limits = c(0, NA), expand = expansion(0, 0.6))
                     )
     } else { ## rectangular
         obj <- coord_cartesian(clip = 'off')